requirements.txt
aiogram-3.15.0
pydantic_settings-2.7.0
psutil
mysql-connector-python-9.1.0